models fixed

+++ app/models.py
@@ -7,7 +7,7 @@ class Invoice(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     currency = db.Column(db.String(3))
     description = db.Column(db.String(120))
-    amount = db.Column(db.Integer)
+    amount = db.Column(db.Float)
     created = db.Column(db.DateTime, default=datetime.utcnow)
 
     def __init__(self, amount, currency, description=''):
